refactor: 패키지 이동 & 메서드 분리
