+++ app/src/main/java/woowacourse/shopping/di/RepositoryBinder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di.repository
+package woowacourse.shopping.di
 
 import com.woowa.di.component.InstallIn
 import com.woowa.di.singleton.SingletonComponent
@@ -9,11 +9,11 @@ import woowacourse.shopping.model.repository.CartRepository
 import woowacourse.shopping.model.repository.ProductRepository
 
 @InstallIn(ViewModelComponent::class)
-class RepositoryBinder {
+class ViewModelRepositoryBinder {
     fun provideProductRepository(): ProductRepository = ProductRepositoryImpl()
 }
 
 @InstallIn(SingletonComponent::class)
-class RepositoryBinder2 {
+class SingletonRepositoryBinder {
     fun provideCartRepository(): CartRepository = CartRepositoryImpl()
 }
