+++ di/src/main/java/com/woowa/di/viewmodel/RemoveField.kt
@@ -1,0 +1,19 @@
+package com.woowa.di.viewmodel
+
+import androidx.lifecycle.ViewModel
+import com.woowa.di.findQualifierClassOrNull
+import javax.inject.Inject
+import kotlin.reflect.jvm.kotlinProperty
+
+inline fun <reified T : ViewModel> removeInstancesOnCleared(instance: T) {
+    instance.addCloseable {
+        T::class.java.declaredFields.onEach { field ->
+            field.isAccessible = true
+        }.filter { it.isAnnotationPresent(Inject::class.java) }.forEach { field ->
+            ViewModelComponentManager.deleteDIInstance(
+                field.type.kotlin,
+                field.kotlinProperty?.findQualifierClassOrNull(),
+            )
+        }
+    }
+}
